增加XML模型读取功能;修改config的bug;修改SerialLink与Link的指针保存方式;将mlab测试文件中的读取机器人部分改为新方法

+++ src/trajectory/SequenceInterpolator.h
@@ -101,7 +101,7 @@ public:
 	T dx(double t) const
 	{
 		int i=0;
-		for (; i<_interpolatorSequence.size(); i++)
+		for (; i<(int)_interpolatorSequence.size(); i++)
 		{
 			if (_timeSequence[i] > t)
 				break;
@@ -115,7 +115,7 @@ public:
 	T ddx(double t) const
 	{
 		int i=0;
-		for (; i<_interpolatorSequence.size(); i++)
+		for (; i<(int)_interpolatorSequence.size(); i++)
 		{
 			if (_timeSequence[i] > t)
 				break;
